rockchip: add workaround disabling jtag/sdmmc autoswitching
rk3288 SoCs have a function to automatically switch between jtag/sdmmc pinmux
settings depending on the card state. This collides with a lot of assumptions.

For one the mmc subsystem does not care about the state of jtag, whose clocks
may very well be disabled. It also only works when using the internal card-
detect mechanism and breaks horribly when using either the main card-detect
via the slot-gpio function or any other pin.

Leaving the switching on can result in mmc timeouts, rcu stalls thus hanging
the system. Therefore disable it in all cases, as we expect the devicetree to
explicitly select either mmc or jtag pinmuxes anyway.

Coreboot for rk3288 already disables this function, so this should also move
into uboot eventually.