pinky: reenable power-button
In contrary to the ChomeOS kernel commit 9426e34ce111 there is at least one
pinky still living - mine ;-) .

The rest of 9426e34ce111 is of course right, the majority of boards should
not need to override the setting of one single board. Therefore add the
necessary overrides to the pinky board file and make the button work again.